Replace monolithic V4.1 tests with compact component checks
