+++ experimental/lite/tests/unit/deepseek_v41/test_engram.py
@@ -1,0 +1,34 @@
+# Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+import pytest
+import torch
+from megatron.lite.model.deepseek_v41.lite import engram
+from megatron.lite.primitive.quantization import ds41_fp8
+
+
+@pytest.mark.parametrize('trainable', [False, True])
+def test_v41_engram_storage_and_reset(trainable):
+    encoded = ds41_fp8.quantize_swa(torch.linspace(-2, 2, 96).reshape(3, 32))
+    table = engram.EngramTable(encoded.values, encoded.scale, trainable=trainable)
+    ids = torch.tensor([[0, 0, 2]])
+    original = table(ids).float()
+    table.bfloat16()
+    assert table.weight.dtype == torch.float8_e4m3fn
+    assert table.scale.dtype == torch.float8_e8m0fnu
+    if trainable:
+        table(ids).float().sum().backward()
+        assert table.master.dtype == torch.float32
+        torch.testing.assert_close(
+            table.master.grad, torch.tensor([2.0, 0.0, 1.0])[:, None].expand(3, 32)
+        )
+        with torch.no_grad():
+            table.master.mul_(16)
+        table.refresh_storage()
+        torch.testing.assert_close(table(ids).float(), original * 16)
+    else:
+        assert not list(table.parameters()) and 'master' not in table.state_dict()
+    hasher = engram.NgramHash(
+        [0, 1, 2, 3], 0, torch.tensor([[3, 5, 7]]), torch.tensor([[[11], [13]]])
+    )
+    hashed = hasher(torch.tensor([[1, 2, 3]]), cu_seqlens=torch.tensor([0, 2, 3]))
+    assert hashed.tolist() == [[[[3, 14]], [[3, 14]], [[9, 20]]]]
+
